cloudfunctions/requirements.txt
google-cloud-dlp
google-cloud-pubsub
google-cloud-bigquery
requests
re2
